requirements.txt
streamlit
python-dotenv
google-generativeai
requests
streamlit-lottie
pytz

-e .

